move ramp to separate component and add transak to top up page
